+++ packages/sources/alphavantage-test/CHANGELOG.md
@@ -1,1 +1,11 @@
 # @chainlink/alphavantage-test-adapter
+
+## 1.0.0
+
+### Major Changes
+
+- 9116aa423: Add alphavantage EA v3
+
+### Patch Changes
+
+- 31af84e69: Bumped framework version