+++ packages/sources/cryptoapis/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/cryptoapis-adapter
 
+## 1.2.30
+
+### Patch Changes
+
+- Updated dependencies [3dc13a0bc]
+  - @chainlink/ea-bootstrap@2.27.0
+  - @chainlink/ea-factories@1.2.29
+  - @chainlink/ea-test-helpers@1.4.2
+
 ## 1.2.29
 
 ### Patch Changes