+++ packages/sources/dns-query/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/dns-query-adapter
 
+## 1.6.22
+
+### Patch Changes
+
+- Updated dependencies [3dc13a0bc]
+  - @chainlink/ea-bootstrap@2.27.0
+  - @chainlink/ea-test-helpers@1.4.2
+
 ## 1.6.21
 
 ### Patch Changes