+++ packages/sources/gemini/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/gemini-adapter
 
+## 2.2.30
+
+### Patch Changes
+
+- Updated dependencies [3dc13a0bc]
+  - @chainlink/ea-bootstrap@2.27.0
+  - @chainlink/ea-test-helpers@1.4.2
+
 ## 2.2.29
 
 ### Patch Changes