+++ packages/sources/iex-cloud/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/iex-cloud-adapter
 
+## 2.0.0
+
+### Major Changes
+
+- 03910e868: Upgraded adapter from v2 to v3 framework
+
+### Patch Changes
+
+- 31af84e69: Bumped framework version
+
 ## 1.2.29
 
 ### Patch Changes